Merge Dev v4

+++ routes/users.js
@@ -31,7 +31,6 @@ router.get('/', passport.authenticate('jwt', { session: false }), async (ctx, ne
 router.get('/:id', passport.authenticate('jwt', { session: false }), async (ctx, next) => {
 	try{
 		let user = await ctx.db.User.findAll({where: {id: ctx.params.id}});
-
 		if (user.length === 0) {
 			throw new Error(`There's no user under id: ${ctx.params.id}`);
 		} else {
@@ -46,16 +45,21 @@ router.get('/:id', passport.authenticate('jwt', { session: false }), async (ctx,
 // Get user projects by id
 router.get('/:id/projects', passport.authenticate('jwt', { session: false }), async (ctx, next) => {
 	try{
-		let userProjects = await ctx.db.Project.findAll({where: {userId: ctx.params.id}});
-
-		if (userProjects.length === 0) {
-			throw new Error(`User with id: ${ctx.params.id} has no projects`);
+		const userProjects = await ctx.db.User.findByPk(ctx.params.id, {
+			include: [ctx.db.Project]
+		});
+		if (!userProjects) {
+			throw new Error(`User with id: ${ctx.params.id} does not exist`);
 		} else {
-			ctx.body = userProjects;
-			next();
+			if (userProjects.Projects.length) {
+				ctx.body = userProjects.Projects;
+				next();
+			} else {
+				throw new Error(`User with id: ${ctx.params.id} has no projects`);
+			}
 		}
 	}catch (ValidationError) {
-		ctx.throw(404, `${ValidationError}`);
+		ctx.throw(400, `${ValidationError}`);
 	}
 });
 
@@ -83,6 +87,7 @@ router.post('/register', async (ctx) => {
 			ctx.body = {user, token};
 		}
 	} catch (ValidationError) {
+		console.log(ValidationError);
 		ctx.throw(400, 'Couldn\'t add the new user');
 	}
 });
@@ -117,27 +122,69 @@ router.post('/login', async (ctx) => {
 // Delete a user by id
 router.delete('/delete/:id', passport.authenticate('jwt', { session: false }), async (ctx) => {
 	try {
-		const deleted = await ctx.db.User.destroy({where: {id: ctx.params.id}});
-		if (deleted > 0) {
-			ctx.response.status = 200;
-			ctx.body = `User ${ctx.params.id} deleted`;
-		} else {
+		let old = await ctx.db.User.findOne({where: {id: ctx.params.id}});
+		if (old){
+			let user = await ctx.db.User.findOne({where: {email: 'deleted@uc.cl'}});
+			let projects = await ctx.db.Project.findAll({where: {userId: ctx.params.id}});
+			let fundings = await ctx.db.Funding.findAll({where: {userId: ctx.params.id}});
+
+			if (projects.length > 0) {
+				projects.map(async (proj) => {
+					return await ctx.db.Project.update({userId: user.dataValues.id, currentState: 'deleted'}, {where: {id: proj.dataValues.id}});
+				});
+			}
+
+			if (fundings.length > 0) {
+				fundings.map(async (fin) => {
+					return await ctx.db.Funding.update({userId: user.dataValues.id}, {where: {id: fin.dataValues.id}});
+				});
+			}
+			
+			try {				
+				let deleted = await ctx.db.User.destroy({where: {id: ctx.params.id}});
+				if (deleted > 0) {
+					ctx.response.status = 200;
+					ctx.body = `User ${ctx.params.id} deleted`;
+				}
+			} finally {
+				await ctx.db.User.destroy({where: {id: ctx.params.id}});
+			}
+		}else {
 			throw new Error('User not found');
 		}
 	} catch (ValidationError) {
-		ctx.throw(400, `${ValidationError}`);
+		console.log(ValidationError.message);
+		let old = await ctx.db.User.findOne({where: {id: ctx.params.id}});
+		if (old){
+			ctx.throw(400, ValidationError);
+		}else{
+			if (ValidationError.message === 'User not found'){
+				ctx.throw(400, ValidationError);
+			}else{
+				console.log('desde catch');
+				ctx.throw(200, `User ${ctx.params.id} deleted`);
+			}
+		}
 	}
 });
-
 
 // Edit user by id
 router.put('/:id', passport.authenticate('jwt', { session: false }), async (ctx) => {
 	try{
 		var data = ctx.request.body;
+
+		if (data.password) {
+			const salt = bcrypt.genSaltSync();
+			const hash = bcrypt.hashSync(data.password, salt);
+			data.password = hash;
+		}
+
 		const update = await ctx.db.User.update(data, {where: {id: ctx.params.id}});
+
 		if (update > 0) {
+			const updated_user = await ctx.db.User.findOne({where: {id: ctx.params.id}});
+			ctx.body = {user: updated_user};
 			ctx.response.status = 200;
-			ctx.body = 'User updated';
 		} else {
 			throw new Error('Something is wrong check the parameters');
 		}
